+++ MacC-GoldenRatio/MacC-GoldenRatio/Sources/Models/ViewModels/MyHomeViewModel.swift
@@ -18,7 +18,6 @@ class MyHomeViewModel {
 	private var myUID = Auth.auth().currentUser?.uid ?? ""
 	
 	let diaryCollectionViewModel = DiaryCollectionViewModel()
-	let mapViewModel = MapViewModel()
 	let albumCollectionViewModel = AlbumCollectionViewModel()
 	
 	var isEqual = false
@@ -51,10 +50,6 @@ class MyHomeViewModel {
 			.bind(to: diaryCollectionViewModel.collectionDiaryData)
 			.disposed(by: disposeBag)
 		
-		diaryValue
-			.map(getFirstDiaryData)
-			.bind(to: mapViewModel.mapDiaryData)
-			.disposed(by: disposeBag)
 	}
 	
 	func getDiaryValue(_ result: Result<[Diary], Error>) -> [Diary] {
@@ -70,14 +65,6 @@ class MyHomeViewModel {
 		}
 		
 		return [DiarySection(header: "다이어리", items: value)]
-	}
-	
-	func getFirstDiaryData(_ value: [Diary]?) -> Diary {
-		guard let value = value else {
-			return Diary(diaryUUID: "", diaryName: "", diaryLocation: Location(locationName: "", locationAddress: "", locationCoordinate: [], locationCategory: nil), diaryStartDate: "", diaryEndDate: "", diaryCover: "")
-		}
-		
-		return value.first ?? Diary(diaryUUID: "", diaryName: "", diaryLocation: Location(locationName: "", locationAddress: "", locationCoordinate: [], locationCategory: nil), diaryStartDate: "", diaryEndDate: "", diaryCover: "")
 	}
 	
 	func isDiaryCodeEqualTo(_ diaryUUID: String) {
