+++ MacC-GoldenRatio/MacC-GoldenRatio/Sources/ViewControllers/Album/MyAlbumPhotoViewController.swift
